rename modules to group
