+++ services/web/server/src/simcore_service_webserver/payments/api.py
@@ -1,8 +1,3 @@
-from ._api import (
-    cancel_payment_to_wallet,
-    init_creation_of_wallet_payment,
-    list_user_payments_page,
-)
 from ._autorecharge_api import (
     get_wallet_payment_autorecharge,
     replace_wallet_payment_autorecharge,
@@ -13,6 +8,11 @@ from ._methods_api import (
     get_wallet_payment_method,
     init_creation_of_wallet_payment_method,
     list_wallet_payment_methods,
+)
+from ._onetime_api import (
+    cancel_payment_to_wallet,
+    init_creation_of_wallet_payment,
+    list_user_payments_page,
 )
 
 __all__: tuple[str, ...] = (
